fix(store): set loose option default value to false (#172)
BREAKING CHANGE: From now, for external storage, the result of the list action will not invalidate when the model instance is changed. The items within the list are still updated, but deletion or creating a model instance won't update the list unless you set the storage `loose` option to `true`. 

However, the default value change only applies to listing enumerable models with external storage, which can be changed by the user. For most of the cases, you should ok, without additional change in the code.