fix(types): typing the "store" descriptor
I qualified the "Model" type to be "NonArrayObject" so that the store() descriptor overloads work correctly.

I defined the "ModelInstance" of the store<ModelInstance>() descriptor to be "NonArrayObject" so that the store() handle overloads work correctly.

Added the "loose" option for the "store" descriptor of the "Enumerable Listing" type.

Added static type tests for the store() descriptor.